- moved `ClipboardManager` to `Core` project - renamed `ClipboardManager` to `ClipboardService` - moved interface `IClipboardService` to separate file
